bug 374916: [model explorer] moving a model element in another model does not move associated tables https://bugs.eclipse.org/bugs/show_bug.cgi?id=374916 -modified the ElementEditHelperAdvice.getAfterMoveCommand method so that the returned command takes care of moving the tables . -added a method in ElementUtil to fetch all the descendants of an element that are papyrus table instances.
Contributed by Olivier Mélois

UNCONFIRMED - bug 375822: [TABLE] Papyrus Table Instances are attached to the wrong resource when created in a controlled package. 
https://bugs.eclipse.org/bugs/show_bug.cgi?id=375822

Changed the resources to which the table instances are attached when the table is created, so that they are saved in the right .di file. 

Contributed by Olivier Mélois