UNCONFIRMED - bug 376703: [Block Definition Diagram] Blocks should have a compartment to display flow properties. https://bugs.eclipse.org/bugs/show_bug.cgi?id=376703
Enabled the usage of the flow properties compartment which was already existing. 
Added a listener that triggers on stereotype additions/removal. It makes sure that the properties are in the right compartments, as flow properties should not be displayed in the "property" compartment, and vice versa. 

Contributed by Olivier Mélois