Bug 425215: Duplicate #include directives
When class uses more than one type from the AnsiCLibrary package,
multiple, identical #include directives are generated.

I've created a utility rule to generate #include directives.  The
parameter to the rule is a string containing the path of the file to be
included.  The rule also fixes a problem where #include's were
specifying system includes (angle-bracket strings) instead of user
includes (quoted strings).

Include directives are now generated in two steps.  In the first step,
all dependencies are examined to create an ordered set of strings with
the include paths.  In the second step the new utility rule is used to
generate the directive.

Change-Id: I11b784f1def4875ecf2c40cae7c3995d7666ef48
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>