Bug 433550: Automated Test Suite for C++ codegen
This is a test suite to confirm basic functionality of the C++ code
generator.  The project includes a model as well as the copies of the
files that are expected to be generated.

The junit test case programmatically runs the generator on each element
and then checks for differences from what is expected.

This needs a minor change in the basic generation code so that it can be
run in a headless fashion.  The change causes the generator to avoid
opening a confirmation dialog on error if a certain System property is
set.

The comparison utility makes allowances for different formatting so that
the expected code is not tied to the formatter that happens to be
selected in the user's workspace.

If a difference is detected, then the individual junit test case is
aborted.

Phase 1 of the test suite (this commit) includes:

11 tests
o   9 Classes (headers, bodies)
o   2 namespace headers

9 class tests covering generation of:
o   Classes
    §  Inheritance
    §  Abstract
    §  Reusability as type in other classes and operations
    §  Default constructor

o   Operations
    §  Empty and defined method bodies
    §  ANSIC return types
    §  Pointer and Reference parameters (in, out, inout) and return
       types
    §  Different classifiers (const, volatile, extern)
    §  Different visibilities
    §  Virtual
    §  Pure virtual
    §  Operator overloading
    §  Polymorphism
    §  Constructor, destructor

o   Attributes
    §  Default values
    §  Static values
    §  Const
    §  Different storage classifiers (volatile, register, extern)
    §  Strings
    §  Char, double, int, void, float
    §  ANSIC types

o   Includes
    §  Include statements
    §  Global definitions

Namespace header tests covering:
o   Packages
    §  Namespace headers
    §  Folder generation

Change-Id: If22a8d3e8e3788f600288c183ac225468db4926a
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>