433310: The Repair Stereotypes functionality should improve the 'Apply Profile' dialog https://bugs.eclipse.org/bugs/show_bug.cgi?id=433310
Improve resolution and presentation of progress metering in the repair actions and present a progress-monitor appears in the dialog (the busyCursorWhile API doesn't show one because the Stereotype Repair dialog is modal).

Includes a new test plug-in for the oep.infra.tools plug-in, which has a new UIUtil class that is JUnit-tested.