422257: [Performances] Memory leaks https://bugs.eclipse.org/bugs/show_bug.cgi?id=422257
Updates to CDO Integration component and tests to fix memory leaks.  This is a work-in-progress.

Especially noteworthy is the new "unload" implementation for CDO views, which clears the adapters (especially UML CacheAdapter) from all CDOObjects in the view when closing it, in lieu of relying on Resource::unload() which CDOResource does not implement.