433371: [Model Explorer] Deleting a diagram keeps its tab open https://bugs.eclipse.org/bugs/show_bug.cgi?id=433371
Rework the DeletePageAdvice as strictly a delete-dependents advice that precomputes the actual PageRefs to delete from the SashModel, so that it doesn't matter that the references from the PageRefs to the deleted diagrams are already removed by the time the advice command executes.

Includes a JUnit integration test to trap future regressions.