434635: [Model Explorer] Preference for "Link with Editor" https://bugs.eclipse.org/bugs/show_bug.cgi?id=434635
Implement memento-based persistence for the link-with-editor and alphabetical-sorting toggle buttons in the Model Explorer tool bar.  Fix the initial calculation of the toggle state of the advance view tool bar item and implement calculation of the toggle state of the show-diagrams tool bar item.

All of these toggle states are shared across all Model Explorer pages (the user perceives the Model Explorer as one view, not a bunch of pages).