417409: [Performances - Properties view] Delay in UI when reorganizing diagram layout. https://bugs.eclipse.org/bugs/show_bug.cgi?id=417409
Make property sheet views reusable, with updating of the bound selection when the selection changes to another element that shows the same views.  This employs new capability of the DataSource to update the selection that it encapsulates, pushing the new selection into the ModelElements that it creates, using a new delegating observable framework.

Property sheet controls are re-used on a per-tab basis.

Because of the new delegation pattern introduced here, we need to be able to ensure that delegate observables are disposed of when they are no longer needed.  This includes not only the delegates of the new DelegatingObservables, but also the delegates of MultipleObservableValue and similar aggregates.  As these delegates can be shared amongst multiple wrappers of different kinds, we use a simple reference counting scheme to ensure that observables are not disposed while they are still in use.  This averts the exceptions discovered in multi-observable (multiple selection) scenarios on a previous iteration of this patch set.

Change-Id: Ide8f3fcea4228083a68bc9d5d39dc5a50217af62