415639: [Resource Loading] Unloading a controlled unit hangs the workbench https://bugs.eclipse.org/bugs/show_bug.cgi?id=415639
Fix ConcurrentModificationException that bombs controlled unit unloading and implement proper progress-monitoring UI in the LoadingUtils.

Also fix a problem in which newly loaded resources appear to be read-only because the ControlledResourcesTracker doesn't detect that there a cross-resource containment proxy has been resolved.

Do not permit unloading of a resource that is dirty.

Change-Id: I1c2242aa7a346209faf586128305f7ed7de06584