405391: [Moka] Moka shall provide an implementation of the fUML Execution Model https://bugs.eclipse.org/bugs/show_bug.cgi?id=405391
Some issues have been noticed when a model is executed for the first
time in a session (not always reproducible).
This problem seems to be related to the fact that, the first time
AnimationUtils.getDiagrams(EObject) is called, it goes through a long
time cross reference resolution process. This may be the cause of the
freezes.
To avoid this issue, this modification forces cross resolutions before
actually starting execution.