Bug 436368: Add missing CDT dependency.
The CDT New project wizard is used to create an output project.  That
wizard is implemented in cdt.ui, but uses dependency injection to
execute code from managedbuilder.ui.  This change adds that bundle
dependency so that binary installations will work properly.

I'm not able to test the buckminster build and install, but I've
simulated this by running with and without the managedbuilder.ui plugin.

Change-Id: Ie28697ecd6744a5222524966d726941b54f96d7d
Signed-off-by: Andrew Eidsness <andrewe@jfront.com>