437052: [CDO] Support resource modification tracking for non-CDO resources https://bugs.eclipse.org/bugs/show_bug.cgi?id=437052
Restore the resource modification-tracking-based strategy in the
ModelSet's save logic.  IModel implementations delegate the question of
whether a resource needs to be saved to the ModelSet via a new
"shouldSave(Resource) : boolean" API.

The ModelSet, in turn, delegates the question of whether a resource
needs to be saved to its ProxyModificationTrackingAdapter, which latter
is now specialized by the CDOAwareModelSet to take into account the save
semantics of CDOResources.  For a ModelSet that is not tracking resource
modifications, and therefore does not have this adapter, it simply
assumes that any resource that can be saved needs to be saved.

Change-Id: Ia651b38f09d7dade435a8738b8a4a7f25e50af1b