405391: [Moka] Moka shall provide an implementation of the fUML Execution Model https://bugs.eclipse.org/bugs/show_bug.cgi?id=405391
- Surrounded executions with try catch. This is to ensure cleaner
termination when exceptions occur in the execution engines.
For the fUML, PSCS, and async execution engine, the execution terminates
if an exception occurs in the main execution thread.
In addition, for the async engine, if an execution occurs in the
execution thread of an active object, this thread terminantes, but the
other threads go on.