437217: [Editors] In-place reloading of model resources in the editors https://bugs.eclipse.org/bugs/show_bug.cgi?id=437217
In situ editor reloading.  Introduces an IReloadableEditor adapter
protocol with an implementation in the CoreMultiDiagramEditor that implements internal destruction of the ServicesRegistry and nested editors.  Some refactoring of the initialization and disposal code in the editor class hierarchy and dependencies facilitates reuse of init/dispose code in the reload scenario.  The re-loading of an editor is deferred until it is next activated, unless it is already the active editor (can happen when "Save All" is invoked).

Editor re-load notifications to dependent views like Model Explorer and Outline.  A new listener protocol informs dependents before and after reload so that they may properly dispose of obsolete state and re-initialize when the editor is reloaded.  Also ensure that an editor is only reloaded once when some resource that it depends on has changed, not once for each resource.

State restoration tokens.  Re-load listeners can insert tokens into the re-load event that capture state to be restored after the re-load. Listeners retrieve and apply these tokens after the editor re-loads itself.  Current state restoration includes:

- tree node expansion and selection state in the Model Explorer view
- diagram outline view: which presentation (tree or overview thumbnail) is active
- which workbench part is active, such that the correct selection is
reflected in views such as Model Explorer, Outline, and Properties
- current active diagram in the re-loaded editor
- edit-part selections in all diagrams
- selection (columns and rows, not individual cells) in table editors
- palettes in each diagram (or palette pages when the Palette View is open):
  * active tool
  * pinnable stack tool selection
  * drawer expansion state
  * drawer scroll position

The Palette View support incidentally fixes loss of palette state when switching between Papyrus editors, caused by the
PapyrusPaletteSynchronizer.

JUnit regression tests for various aspects of editor re-load.

Includes a fix for an NPE in the Validation View's content provider that occurs in several tests when an editor is closed or re-loaded.  Also support for tests that need to load more than one test-fixture model and/or open more than one editor.

Change-Id: Ic0f654ab138d3e091f81f1e9159bcca80d6bb0a5