Bug 422532: Create Volatile stereotype
This fixes a problem where volatile was included in the EStorageClass
enum.  This does not match the C++ language spec and caused problems
(see Bugs 425594, 425597, and 425598) when the StorageClass stereotype
was applied to invalid language elements.

I've removed volatile from that enum and removed the extension from
StorageClass to Parameter and Property.  I've also introduced a new
Volatile stereotype and created new extensions that mirror the ones for
Const.

I've updated the code generator to use this stereotype in the same
places that Const is currently checked.  I also fixed a problem where
const (and now volatile) was generated for static member functions.

I've updated the TestCDTIntegration model but not the other ones yet.  I
would like to make sure this change is OK before making changes to the
other two test models.

When generating the profile I found that uml.ecore and types.ecore were
not found.  I've changed the C_Cpp.ecore file to use the platform: URIs
for these references but I'm not convinced that is the right solution.

I've also updated the namespace URI from /1 to /2.  This change was
needed to get existing models to reapply the profile.  I think that is
the expected method.

Change-Id: Ic572ac6a549d2c2421632a2b66414066d366b608
Signed-off-by: Andrew Eidsness <andrewe@jfront.com>