437217: Restore normal control-mode strategy for tests https://bugs.eclipse.org/bugs/show_bug.cgi?id=437217
Ensure that tests that operate on control mode strategies restore the strategy that was in effect before they run (usually the default, load-everything strategy) when they finish, to avoid interfering with the loading of resources in subsequent tests in the suite execution.

The tests for in-place re-loading of editors (bug 437217) were failing because resources from library dependencies weren't being loaded and therefore weren't triggering re-load when changes were saved by other editors.