440197: The "Apply" action in "Repair Stereotypes" dialog doesn't work for the first click when more than 2 (incl. 2) profiles involved. https://bugs.eclipse.org/bugs/show_bug.cgi?id=440197
Fix display of repairable applied stereotypes in diagrams. This includes updating the responsible edit-part to schedule deletion of the stereotype visuals (or refreshing to pick up repair results) after the completion of the repair user interaction.

This deferral until repair is completed was actually broken: deferred actions would be run prematurely in cases where repair loads additional UML resources that weren't previously loaded. So, that is fixed in order to let the diagram properly defer treatment of broken stereotype visuals.

Change-Id: I5510ce304ef7c13d4ca4449073e274298608f013