437217: [Editors] In-place reloading of model resources in the editors https://bugs.eclipse.org/bugs/show_bug.cgi?id=437217
Outline pages need to be treated the same as property pages as far as editor re-loading and the editor lifecycle is concerned.  They must only be forgotten at disposal of the editor because they track creation/destruction of the editor's nested pages and don't depend, themselves, on whether the editor has been re-loaded.