437217: [Editors] In-place reloading of model resources in the editors https://bugs.eclipse.org/bugs/show_bug.cgi?id=437217
Implement deferral of the prompt to re-load until the editor is activated, in addition to the actual re-loading.  Also defer prompting and closing of editors for deleted resources until activation, for consistency and to avoid confusing dialogs that don't appear to have any effect or are inappropriate (esp. in Save All scenarios).