441857: [Performances - Model Explorer] Severe performance problems for larger models https://bugs.eclipse.org/bugs/show_bug.cgi?id=441857
Don't refresh the entire model explorer every time the selection in a
diagram editor changes. Moreover, let requests for synchronous refresh
merely hurry up a pending asynchronous refresh (if any). This works on
the assumption that only changes in the resource set and/or read-only
state should invalidate the explorer's presentation, which is supported
by the fact that only these triggers schedule an asynchronous refresh
in the first place.

Change-Id: I089723942e59994280dc4dd8eceafc5f35c13fac