[Releng] Fix setup model issues and Git project layout
The new JUnit Framework plug-in must always be imported by the setup model as the JUnit Utils plug-in is.

The releng/ folder also is a project, which needs to be accounted for.

The oep.serviceregistry.tests plug-in's storage folder does not match the bundle symbolic name, which causes Oomph project import to fail.

New test plug-ins/fragments need to have common code style settings propagated.