443417: [Properties view] Gradient editor not refreshed when switching selection https://bugs.eclipse.org/bugs/show_bug.cgi?id=443417
The ViewEditor class in the properties-view framework implements what amounts to
a nested property sheet.  Accordingly, when its containing property editors are
reused for a new selection, it also needs to reuse its contents for the new selection.
This requires keeping the same DisplayEngine to be reused and reusing the nested
XWTSections that it manages.