441227: [Properties view] Create New Property view tabs not working https://bugs.eclipse.org/bugs/show_bug.cgi?id=441227
It is not currently feasible to apply stereotypes to elements that are
new and not yet
attached to the model.  So, define a new properties-view constraint to
test for the
selection being attached to a resource and constrain the SingleElement
section of
the Profile tab accordingly.

In testing, an NPE was encountered when a disposed Stereotype
Applications
table responded to changes in the stereotypes applied to an element in
the
dialog.  That is fixed by detecting the disposed widget and removing the
callback.

Change-Id: I5116c00a881306cffd086c6c5fce0ad35733aaff