444092: [Properties view] Refresh issues for stereotype properties https://bugs.eclipse.org/bugs/show_bug.cgi?id=444092
The source of a StereotypeModelElement is the application (instance) of the stereotype applied.  So, when updating the model-element for a change in selection, the new source must be the new UML element's application of the stereotype, not the new UML element.  Moreover, we must be careful that the StereotypeModelElement tracks the stereotype indicated by the DataContext, which may be a superstereotype (possibly abstract or otherwise not actually applicable) of the stereotype actually applied to any given UML element.

Change-Id: I8e7084eca245a95f0af11bc1d8a4a7e0e813dd00