444227: Property view broken for associations https://bugs.eclipse.org/bugs/show_bug.cgi?id=444227
The ViewEditor must be able to repeat sections for list values, such as the multiple ends of an association.  To that end, the DisplayEngine now supports proxy sections that combine the same section instance that needs to be repeated with an arbitrary discriminator that distinguishes the repeated occurrences.  This allows the display engine to reuse controls for the same repetition of the same section.

Also, because multiple selections are presented as observables that are dynamic proxies wrapping a MultipleObservableValue, the ViewEditor cannot rely on that class's API to access the multiple objects that were selected.  The API required by the ViewEditor is factored out into an interface that the proxy then automatically implements, as usual in the delegating observables mechanism.

Change-Id: I88345c23f898100bd109bab2eccfd60d4b098323