449474: Do not load ImageRegistry in headless mode.
The test PlatformUI.isWorkbenchRunning() will allow to not invoke the
eclipse ui plugin if eclipse is an headless mode.

Bug: 449474

I, Axel Richard, wrote 100% of the code I've provided.
This code contains no cryptography.
I have the right to contribute the code to Eclipse.
I contribute the content under the EPL.

Change-Id: I0d3154f6d4a979a416fc39abee5fa472c37c5825
Signed-off-by: Axel Richard <axel.richard@obeo.fr>