399859: [Profile Applications] Papyrus shall enable to manage profile applications in separate files https://bugs.eclipse.org/bugs/show_bug.cgi?id=399859
Ensure that we don't get stack-overflow on recursion in the read-only handler for decorator models in the case that a decorator model creates an externalized profile application for one of its own decorator packages.  Also ensure that stereotype repair doesn't accidentally create such an externalized profile application for a decorator package when repairing a decorator model.