454536: [Properties View] Impossible to display the properties pages after reopening of a project https://bugs.eclipse.org/bugs/show_bug.cgi?id=454536
Ensure that the current selection in the Property sheet is forgotten when closing the model if the selection is from that model. Also ensure that the property sheet always updates to show the current Model Explorer selection when the explorer view is activated (otherwise it may still be showing what was relevant for a previous model editor).

Furthermore, when re-using a data source in the property sheet, if that data source contains unloaded objects, then don't reuse it, to ensure that widgets aren't re-used that remember the previous editor's resource-set and other context.