436666: "Stereotype repair" pop-up for already upgraded model https://bugs.eclipse.org/bugs/show_bug.cgi?id=436666
Detect orphaned (dangling) stereotype instances and suggest to delete them.
Orphans are treated as a separate group from properly attached stereotype
instances of the same schema (broken or not) because they cannot be repaired
by profile application; they are not stereotype applications because they have
no base elements.

Also update some test models that are opened in editors to remove orphaned stereotype instances (which cause tests to hang on the repair dialog).

Conflicts:
	extraplugins/migration/org.eclipse.papyrus.migration.rsa/src/org/eclipse/papyrus/migration/rsa/transformation/RepairStereotypes.java
	plugins/uml/org.eclipse.papyrus.uml.modelrepair/src/org/eclipse/papyrus/uml/modelrepair/internal/stereotypes/ZombieStereotypesDescriptor.java
	plugins/uml/org.eclipse.papyrus.uml.modelrepair/src/org/eclipse/papyrus/uml/modelrepair/ui/ZombieStereotypesDialog.java