458736: Opening an element/package imported model causes the stereotype repair dialog to repeatedly popup for any of its fragments/submodels https://bugs.eclipse.org/bugs/show_bug.cgi?id=458736
Scan for broken stereotypes asynchronously (on the UI thread before the repair dialog is opened) because the resource being loaded may be a sub-unit that is not yet connected to its parent unit.  Containment proxy resolution will load the resource before attaching its root element to that root's parent, causing stereotypes that depend on profile applications in the parent unit to appear broken when in fact they aren not.